+++ app/Http/Requests/CocktailRequest.php
@@ -46,6 +46,7 @@ class CocktailRequest extends FormRequest
             'ingredients.*.ingredient.substitutes.*.amount' => 'numeric',
             'ingredients.*.ingredient.substitutes.*.amount_max' => 'nullable|numeric',
             'author' => 'nullable|string',
+            'origin_bar' => 'nullable|string',
         ];
     }
 }