+++ app/Http/Resources/CocktailResource.php
@@ -62,6 +62,7 @@ use Illuminate\Http\Resources\Json\JsonResource;
         new OAT\Property(property: 'varieties', type: 'array', items: new OAT\Items(type: CocktailBasicResource::class), description: 'List of varieties of this cocktail'),
         new OAT\Property(property: 'year', type: 'number', example: 2023, description: 'Cocktail recipe year', nullable: true),
         new OAT\Property(property: 'author', type: 'string', example: 'Jerry Thomas', description: 'Historical author of the cocktail recipe', nullable: true),
+        new OAT\Property(property: 'origin_bar', type: 'string', example: 'American Bar, London', description: 'Historical bar or venue where the cocktail was originally created', nullable: true),
     ],
     required: [
         'id',
@@ -142,6 +143,7 @@ class CocktailResource extends JsonResource
             'varieties' => CocktailBasicResource::collection($this->whenLoaded('cocktailVarieties')),
             'year' => $this->year,
             'author' => $this->author,
+            'origin_bar' => $this->origin_bar,
         ];
     }
 }