+++ routes/api.php
@@ -37,6 +37,9 @@ use Kami\Cocktail\Http\Controllers\PriceCategoryController;
 use Kami\Cocktail\Http\Middleware\EnsureRequestHasBarQuery;
 use Kami\Cocktail\Http\Controllers\CocktailMethodController;
 use Kami\Cocktail\Http\Controllers\MemberInventoryController;
+use Kami\Cocktail\Http\Controllers\TasteDescriptorController;
+use Kami\Cocktail\Http\Controllers\IngredientRatingController;
+use Kami\Cocktail\Http\Controllers\IngredientReviewController;
 use Kami\Cocktail\Http\Middleware\AiImageProviderIsConfigured;
 
 /*
@@ -81,7 +84,13 @@ Route::prefix('exports')->group(function () {
     Route::get('/{id}/download', [ExportController::class, 'download'])->name('exports.download');
 });
 
-Route::post('/billing/webhook', WebhookController::class);
+if (config('cashier.webhook_secret')) {
+    Route::post('/billing/webhook', WebhookController::class);
+} else {
+    Route::post('/billing/webhook', function () {
+        abort(403, 'Paddle webhook secret is not configured. Set the PADDLE_WEBHOOK_SECRET environment variable.');
+    });
+}
 
 Route::prefix('public')->group(function () {
     Route::get('/links/cocktails/{publicId}', [Public\CocktailController::class, 'showPublicLinkCocktail']);
@@ -111,7 +120,21 @@ Route::middleware($apiMiddleware)->group(function () {
         Route::get('/{idOrSlug}/cocktails', [IngredientController::class, 'cocktails'])->middleware(['ability:ingredients.read']);
         Route::get('/{idOrSlug}/substitutes', [IngredientController::class, 'substitutes'])->middleware(['ability:ingredients.read']);
         Route::get('/{idOrSlug}/tree', [IngredientController::class, 'tree'])->middleware(['ability:ingredients.read']);
+
+        Route::prefix('/{id}/reviews')->middleware([EnsureRequestHasBarQuery::class])->group(function () {
+            Route::get('/', [IngredientReviewController::class, 'index'])->name('ingredient-reviews.index')->middleware(['ability:ingredients.read']);
+            Route::post('/', [IngredientReviewController::class, 'store'])->name('ingredient-reviews.store')->middleware(['ability:ingredients.write']);
+            Route::put('/{reviewId}', [IngredientReviewController::class, 'update'])->name('ingredient-reviews.update')->middleware(['ability:ingredients.write']);
+            Route::delete('/{reviewId}', [IngredientReviewController::class, 'destroy'])->name('ingredient-reviews.destroy')->middleware(['ability:ingredients.write']);
+        });
+
+        Route::prefix('/{id}/ratings')->middleware([EnsureRequestHasBarQuery::class])->group(function () {
+            Route::post('/', [IngredientRatingController::class, 'rate'])->name('ingredient-ratings.rate')->middleware(['ability:ingredients.write']);
+            Route::delete('/', [IngredientRatingController::class, 'unrate'])->name('ingredient-ratings.unrate')->middleware(['ability:ingredients.write']);
+        });
     });
+
+    Route::get('/taste-descriptors', [TasteDescriptorController::class, 'index'])->middleware([EnsureRequestHasBarQuery::class, 'ability:ingredients.read']);
 
     Route::prefix('cocktails')->group(function () {
         Route::get('/', [CocktailController::class, 'index'])->name('cocktails.index')->middleware([EnsureRequestHasBarQuery::class, 'ability:cocktails.read']);