+++ src/Application/Cocktail/CocktailService.php
@@ -64,6 +64,7 @@ final readonly class CocktailService
             methodId: $request->methodId ? new MethodId($request->methodId) : null,
             variantOf: $variantOf?->getId(),
             author: $request->author,
+            originBar: $request->originBar,
         );
 
         foreach ($request->ingredients as $requestIngredient) {
@@ -140,6 +141,7 @@ final readonly class CocktailService
             variantOf: $variantOf?->getId(),
             year: $request->year,
             author: $request->author,
+            originBar: $request->originBar,
         );
 
         $cocktail->removeAllIngredients();