+++ src/Application/Cocktail/DTO/UpdateCocktail.php
@@ -31,6 +31,7 @@ final readonly class UpdateCocktail
         public ?int $parentCocktailId,
         public ?int $year,
         public ?string $author = null,
+        public ?string $originBar = null,
     ) {
     }
 }